+++ rules.go
@@ -19,7 +19,7 @@ var grammarRules = func(num int) map[string]*Rule {
 	end := num % 10
 
 	return map[string]*Rule{
-		"en,nl,de,es": {
+		"en,nl,de,es,fr": {
 			Zero: num == 0,
 			One:  num == 1,
 			Two:  num == 2,