+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added support for Belarusian language 🇧🇾
 - Added support for Spanish language 🇪🇸
 - Added support for Japanese language 🇯🇵
+- Added support for French language 🇫🇷
 - Code refactoring and improvements for better readability and maintainability
 
 ## v3.1.0 (2024-11-25)